Factor mount utility code out gce_pd volume plugin
