--- pkg/kubelet/volume/gce_pd/mount_util.go
@@ -1,48 +1,0 @@
-/*
-Copyright 2014 Google Inc. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package gce_pd
-
-import (
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/util/mount"
-)
-
-// Examines /proc/mounts to find all other references to the device referenced
-// by mountPath.
-func getMountRefs(mounter mount.Interface, mountPath string) ([]string, error) {
-	mps, err := mounter.List()
-	if err != nil {
-		return nil, err
-	}
-
-	// Find the device name.
-	deviceName := ""
-	for i := range mps {
-		if mps[i].Path == mountPath {
-			deviceName = mps[i].Device
-			break
-		}
-	}
-
-	// Find all references to the device.
-	var refs []string
-	for i := range mps {
-		if mps[i].Device == deviceName && mps[i].Path != mountPath {
-			refs = append(refs, mps[i].Path)
-		}
-	}
-	return refs, nil
-}
