+++ pkg/util/mount/mountpoint_unix.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package gce_pd
+package mount
 
 import (
 	"os"
@@ -26,7 +26,7 @@ import (
 // Determine if a directory is a mountpoint, by comparing the device for the directory
 // with the device for it's parent.  If they are the same, it's not a mountpoint, if they're
 // different, it is.
-func isMountPoint(file string) (bool, error) {
+func IsMountPoint(file string) (bool, error) {
 	stat, err := os.Stat(file)
 	if err != nil {
 		return false, err
